Rename fetch json to fetch proxy; Add docs for fetch proxy;
