+++ src/fetch-proxy.spec.js
@@ -1,10 +1,10 @@
-import fetchJSON from './fetch-json';
+import fetchProxy from './fetch-proxy';
 import { expect } from 'chai';
 import sinon from 'sinon';
 
-describe.only('Fetch JSON', () => {
+describe('Fetch JSON', () => {
   it('returns a function', () => {
-    expect(fetchJSON()).to.be.a('function');
+    expect(fetchProxy()).to.be.a('function');
   });
 
   describe('function', () => {
@@ -21,14 +21,14 @@ describe.only('Fetch JSON', () => {
 
     it('calls fetch with the same arguments as it received', () => {
       fetchFunction.returns(Promise.resolve());
-      fetchJSON(fetchFunction)('abc', 123);
+      fetchProxy(fetchFunction)('abc', 123);
       expect(fetchFunction.calledWith('abc', 123)).to.equal(true);
     });
 
     describe('when fetch gets rejected', () => {
       it('rejects with status 500', (done) => {
         fetchFunction.returns(Promise.reject('error'));
-        fetchJSON(fetchFunction)().catch((err) => {
+        fetchProxy(fetchFunction)().catch((err) => {
           expect(err).to.deep.equal({
             status: 500,
             body: {
@@ -44,7 +44,7 @@ describe.only('Fetch JSON', () => {
       it('rejects with 401', (done) => {
         response.status = 401;
         fetchFunction.returns(Promise.resolve(response));
-        fetchJSON(fetchFunction)().catch((err) => {
+        fetchProxy(fetchFunction)().catch((err) => {
           expect(err).to.deep.equal({
             status: 401,
             body: {
@@ -58,7 +58,7 @@ describe.only('Fetch JSON', () => {
       it('rejects with 404', (done) => {
         response.status = 404;
         fetchFunction.returns(Promise.resolve(response));
-        fetchJSON(fetchFunction)().catch((err) => {
+        fetchProxy(fetchFunction)().catch((err) => {
           expect(err).to.deep.equal({
             status: 404,
             body: {
@@ -72,7 +72,7 @@ describe.only('Fetch JSON', () => {
       it('rejects any other error codes with a default error', (done) => {
         response.status = 300003;
         fetchFunction.returns(Promise.resolve(response));
-        fetchJSON(fetchFunction)().catch((err) => {
+        fetchProxy(fetchFunction)().catch((err) => {
           expect(err).to.deep.equal({
             status: 300003,
             body: {
@@ -87,7 +87,7 @@ describe.only('Fetch JSON', () => {
     describe('when fetch gets resolved with status 200', () => {
       it('wraps the data, parses them into json and resolves', (done) => {
         fetchFunction.returns(Promise.resolve(response));
-        fetchJSON(fetchFunction)().then((res) => {
+        fetchProxy(fetchFunction)().then((res) => {
           expect(res).to.deep.equal({
             status: 200,
             body: 'data'
